+++ src/methods/$find/__tests__/find.spec.ts
@@ -6,53 +6,15 @@
  * More information can be found in CONTRIBUTING.md
  */
 
-import { find, asyncFind, range } from '../../..';
+import { find } from '../../..';
 describe('find', () => {
   it('returns found item', () => {
-    const found = find(item => item === 5, [1, 2, 3, 4, 5, 6]);
-    expect(found).toBe(5);
+    expect(find(item => item === 5, [1, 2, 3, 4, 5, 6])).toBe(5);
   });
   it('returns undefined if no item found', () => {
-    const found = find(_ => false, [1, 2, 3, 4, 5, 6]);
-    expect(found).toBe(undefined);
+    expect(find(_ => false, [1, 2, 3, 4, 5, 6])).toBe(undefined);
   });
-  it('returns notFoundValue if specified and no item found', () => {
-    const found = find(0, _ => false, [1, 2, 3, 4, 5, 6]);
-    expect(found).toBe(0);
-  });
-  it('returns found item from iterable', () => {
-    const found = find(
-      item => item === 5,
-      range({
-        start: 1,
-        end: 7,
-      }),
-    );
-    expect(found).toBe(5);
-  });
-  it('returns undefined if no item found from iterable', () => {
-    const found = find(
-      item => item === 100,
-      range({
-        start: 1,
-        end: 7,
-      }),
-    );
-    expect(found).toBe(undefined);
-  });
-  it('returns filtered iterable (curried version)', () => {
-    const findFive = find(item => item === 5);
-    expect(
-      findFive(
-        range({
-          start: 1,
-          end: 7,
-        }),
-      ),
-    ).toBe(5);
-  });
-  it('returns undefined if passed null', () => {
-    const found = find(item => item, null);
-    expect(found).toBe(undefined);
+  it('returns undefined when iterable is empty', () => {
+    expect(find(item => item, null)).toBe(undefined);
   });
 });